+++ Projects/App/Sources/AppDelegate.swift
@@ -50,18 +50,15 @@ import hGraphQL
         globalPresentableStoreContainer = PresentableStoreContainer()
 
         self.setupSession()
-
-        self.bag += ApolloClient.initAndRegisterClient()
-            .onValue { _ in
-                self.bag += self.window.present(AppJourney.main)
-                UIView.transition(
-                    with: self.window,
-                    duration: 0.3,
-                    options: .transitionCrossDissolve,
-                    animations: {},
-                    completion: { _ in }
-                )
-            }
+        ApolloClient.initAndRegisterClient()
+        self.bag += self.window.present(AppJourney.main)
+        UIView.transition(
+            with: self.window,
+            duration: 0.3,
+            options: .transitionCrossDissolve,
+            animations: {},
+            completion: { _ in }
+        )
     }
 
     func logout() {
@@ -180,10 +177,10 @@ import hGraphQL
         }
         setupAnalyticsAndTracking()
         bag += Localization.Locale.$currentLocale
-            .onValue { [weak self] locale in
+            .onValue { locale in
                 ApplicationState.setPreferredLocale(locale)
                 ApolloClient.acceptLanguageHeader = locale.acceptLanguageHeader
-                self?.bag += ApolloClient.initAndRegisterClient()
+                ApolloClient.initAndRegisterClient()
             }
 
         ApolloClient.bundle = Bundle.main
@@ -276,105 +273,27 @@ import hGraphQL
                 Toasts.shared.displayToast(toast: toast)
             }
         }
-        setupExperiments()
 
         bag += ApplicationContext.shared.$isDemoMode.onValue { value in
             let store: UgglanStore = globalPresentableStoreContainer.get()
             TokenRefresher.shared.isDemoMode = value
             store.send(.setIsDemoMode(to: value))
+            ApolloClient.initAndRegisterClient()
         }
         let store: UgglanStore = globalPresentableStoreContainer.get()
         ApplicationContext.shared.$isDemoMode.value = store.state.isDemoMode
         TokenRefresher.shared.isDemoMode = store.state.isDemoMode
-
+        setupExperiments()
         observeNotificationsSettings()
         return true
     }
 
     private func setupExperiments() {
-        self.bag += ApolloClient.initAndRegisterClient().valueSignal.map { _ in true }.plain()
-            .atValue { _ in
-                self.setupFeatureFlags(onComplete: { success in
-                    DispatchQueue.main.async {
-                        self.bag += self.window.present(AppJourney.main)
-                    }
-                })
+        ApolloClient.initAndRegisterClient()
+        self.setupFeatureFlags(onComplete: { success in
+            DispatchQueue.main.async {
+                self.bag += self.window.present(AppJourney.main)
             }
-    }
-}
-
-extension ApolloClient {
-    public static func initAndRegisterClient() -> Future<Void> {
-        Self.initClients()
-            .onValue { hApollo in
-                Dependencies.shared.add(module: Module { AnalyticsCoordinator() })
-                let paymentService = hPaymentServiceOctopus()
-                let hForeverCodeService = hForeverCodeServiceOctopus()
-                let hCampaignsService = hCampaingsServiceOctopus()
-                let networkClient = NetworkClient()
-                let messagesClient = FetchMessagesClientOctopus()
-                let sendMessage = SendMessagesClientOctopus()
-                let moveFlowService = MoveFlowServiceOctopus()
-                let foreverService = ForeverServiceOctopus()
-                let profileService = ProfileServiceOctopus()
-                let editCoInsuredService = EditCoInsuredServiceOctopus()
-                let homeService = HomeServiceOctopus()
-                let terminateContractsService = TerminateContractsOctopus()
-                Dependencies.shared.add(module: Module { hApollo.octopus })
-                Dependencies.shared.add(module: Module { () -> ChatFileUploaderClient in networkClient })
-                Dependencies.shared.add(module: Module { () -> FetchMessagesClient in messagesClient })
-                Dependencies.shared.add(module: Module { () -> SendMessageClient in sendMessage })
-                let featureFlagsUnleash = FeatureFlagsUnleash(environment: Environment.current)
-                Dependencies.shared.add(module: Module { hApollo.octopus })
-                Dependencies.shared.add(module: Module { () -> FeatureFlags in featureFlagsUnleash })
-                Dependencies.shared.add(
-                    module: Module { () -> TravelInsuranceClient in TravelInsuranceClientOctopus() }
-                )
-
-                switch Environment.current {
-                case .staging:
-                    let hFetchClaimService = FetchClaimServiceOctopus()
-                    Dependencies.shared.add(module: Module { () -> FileUploaderClient in networkClient })
-                    Dependencies.shared.add(module: Module { () -> AdyenService in networkClient })
-                    Dependencies.shared.add(module: Module { () -> hPaymentService in paymentService })
-                    Dependencies.shared.add(module: Module { () -> hForeverCodeService in hForeverCodeService })
-                    Dependencies.shared.add(module: Module { () -> hCampaignsService in hCampaignsService })
-                    Dependencies.shared.add(module: Module { () -> hFetchClaimService in hFetchClaimService })
-                    Dependencies.shared.add(module: Module { () -> hClaimFileUploadService in networkClient })
-                    Dependencies.shared.add(
-                        module: Module { () -> FetchContractsService in FetchContractsServiceOctopus() }
-                    )
-                    Dependencies.shared.add(module: Module { () -> MoveFlowService in moveFlowService })
-                    Dependencies.shared.add(module: Module { () -> ForeverService in foreverService })
-                    Dependencies.shared.add(module: Module { () -> ProfileService in profileService })
-                    Dependencies.shared.add(module: Module { () -> EditCoInsuredService in editCoInsuredService })
-                    Dependencies.shared.add(module: Module { () -> HomeService in homeService })
-                    Dependencies.shared.add(
-                        module: Module { () -> TerminateContractsService in terminateContractsService }
-                    )
-                case .production, .custom:
-                    let hFetchClaimService = FetchClaimServiceOctopus()
-                    Dependencies.shared.add(module: Module { () -> FileUploaderClient in networkClient })
-                    Dependencies.shared.add(module: Module { () -> AdyenService in networkClient })
-                    Dependencies.shared.add(module: Module { () -> hPaymentService in paymentService })
-                    Dependencies.shared.add(module: Module { () -> hForeverCodeService in hForeverCodeService })
-                    Dependencies.shared.add(module: Module { () -> hCampaignsService in hCampaignsService })
-                    Dependencies.shared.add(module: Module { () -> hFetchClaimService in hFetchClaimService })
-                    Dependencies.shared.add(module: Module { () -> hClaimFileUploadService in networkClient })
-                    Dependencies.shared.add(
-                        module: Module { () -> FetchContractsService in FetchContractsServiceOctopus() }
-                    )
-
-                    Dependencies.shared.add(module: Module { () -> MoveFlowService in moveFlowService })
-                    Dependencies.shared.add(module: Module { () -> ForeverService in foreverService })
-                    Dependencies.shared.add(module: Module { () -> ProfileService in profileService })
-                    Dependencies.shared.add(module: Module { () -> EditCoInsuredService in editCoInsuredService })
-                    Dependencies.shared.add(module: Module { () -> HomeService in homeService })
-                    Dependencies.shared.add(
-                        module: Module { () -> TerminateContractsService in terminateContractsService }
-                    )
-                }
-            }
-            .toVoid()
+        })
     }
 }
