+++ Projects/Profile/Sources/Services/ProfileDemoService.swift
@@ -1,6 +1,8 @@
 import Foundation
 
 public class ProfileDemoService: ProfileService {
+    
+    public init() {}
     public func getProfileState() async throws -> (memberData: MemberDetails, partnerData: PartnerData?) {
         return (
             MemberDetails(id: "", firstName: "", lastName: "", phone: "", email: "", hasTravelCertificate: false), nil
