+++ Projects/TravelCertificate/Sources/Services/DemoImplementation/TravelInsuranceClientDemo.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+public struct TravelInsuranceClientDemo: TravelInsuranceClient {
+
+    public init() {}
+
+    public func getSpecifications() async throws -> TravelInsuranceSpecification {
+        throw TravelInsuranceError.missingURL
+    }
+    public func submitForm(dto: TravenInsuranceFormDTO) async throws -> URL {
+        throw TravelInsuranceError.missingURL
+    }
+    public func getList() async throws -> [TravelCertificateModel] {
+        throw TravelInsuranceError.missingURL
+    }
+}
