+++ Projects/hCore/Sources/FeatureFlags/FeatureFlagOctopus.swift
@@ -2,29 +2,6 @@ import Foundation
 import UnleashProxyClientSwift
 import hGraphQL
 
-public protocol FeatureFlags {
-    var loadingExperimentsSuccess: (Bool) -> Void { get set }
-    var isMovingFlowEnabled: Bool { get set }
-    var isEditCoInsuredEnabled: Bool { get set }
-    var isTravelInsuranceEnabled: Bool { get set }
-    var isTerminationFlowEnabled: Bool { get set }
-    var isUpdateNecessary: Bool { get set }
-    var isChatDisabled: Bool { get set }
-    var isPaymentScreenEnabled: Bool { get set }
-    var isCommonClaimEnabled: Bool { get set }
-    var isForeverEnabled: Bool { get set }
-    var paymentType: PaymentType { get set }
-    var isHelpCenterEnabled: Bool { get set }
-
-    func setup(with context: [String: String], onComplete: @escaping (_ success: Bool) -> Void)
-    func updateContext(context: [String: String])
-}
-
-public enum PaymentType {
-    case trustly
-    case adyen
-}
-
 public class FeatureFlagsUnleash: FeatureFlags {
     private var unleashClient: UnleashClient?
     private var environment: Environment
@@ -47,6 +24,7 @@ public class FeatureFlagsUnleash: FeatureFlags {
     public var isForeverEnabled: Bool = false
     public var paymentType: PaymentType = .trustly
     public var isHelpCenterEnabled: Bool = false
+    public var isSubmitClaimEnabled: Bool = true
 
     public func setup(with context: [String: String], onComplete: @escaping (_ success: Bool) -> Void) {
         unleashClient?.unsubscribe(name: "ready")
@@ -163,39 +141,5 @@ public class FeatureFlagsUnleash: FeatureFlags {
             "Feature flag set",
             attributes: ["featureFlags": featureFlags]
         )
-    }
-}
-
-public class FeatureFlagsDemo: FeatureFlags {
-    public init() {}
-
-    public var loadingExperimentsSuccess: (Bool) -> Void = { _ in }
-    public var isMovingFlowEnabled: Bool = false
-    public var isEditCoInsuredEnabled: Bool = false
-    public var isTravelInsuranceEnabled: Bool = false
-    public var isTerminationFlowEnabled: Bool = false
-    public var isUpdateNecessary: Bool = false
-    public var isChatDisabled: Bool = false
-    public var isPaymentScreenEnabled: Bool = false
-    public var isCommonClaimEnabled: Bool = false
-    public var isForeverEnabled: Bool = false
-    public var paymentType: PaymentType = .trustly
-    public var isHelpCenterEnabled: Bool = false
-
-    public func setup(with context: [String: String], onComplete: @escaping (_ success: Bool) -> Void) {
-        loadingExperimentsSuccess = onComplete
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-            self.loadingExperimentsSuccess(true)
-        }
-    }
-
-    public func updateContext(context: [String: String]) {
-    }
-}
-
-extension Dependencies {
-    static public func featureFlags() -> FeatureFlags {
-        let featureFlags: FeatureFlags = shared.resolve()
-        return featureFlags
     }
 }
